Fixed 'add_list_emails'
Previously, due to the use of cURL library to construct and encode URL
query strings 'add_list_emails' only allowed to add a single email
(recipient) to a recipient list at any one time.
Modified 'add_list_emails' to accept an array of recipient information
that gets sent to SendGrid in a format their API accepts allowing the
addition of several emails (recipients) to a recipient list in a single
API call.